+++ one/params.py
@@ -8,7 +8,6 @@ directories, and a separate parameter file for each url containing the client pa
 caches file also sets the default client for when no url is provided.
 
 TODO Rename 'client' kwarg
-FIXME Remove redundant cache_dir param (use directory from client map instead)
 """
 import re
 from iblutil.io import params as iopar
@@ -24,13 +23,11 @@ CACHE_DIR_DEFAULT = str(Path.home() / "Downloads" / "ONE")
 
 def default():
     """Default WebClient parameters"""
-    par = {"ALYX_LOGIN": "intbrainlab",
-           "ALYX_PWD": "international",
-           "ALYX_URL": "https://openalyx.internationalbrainlab.org",
+    par = {"ALYX_URL": "https://openalyx.internationalbrainlab.org",
+           "ALYX_LOGIN": "intbrainlab",
            "HTTP_DATA_SERVER": "https://ibl.flatironinstitute.org",
            "HTTP_DATA_SERVER_LOGIN": None,
-           "HTTP_DATA_SERVER_PWD": None,
-           }
+           "HTTP_DATA_SERVER_PWD": None}
     return iopar.from_dict(par)
 
 
@@ -125,7 +122,20 @@ def setup(client=None, silent=False, make_default=None):
     return cache_map
 
 
-def get(silent=False, client=None):
+def get(client=None, silent=False):
+    """Returns the AlyxClient parameters
+
+    Parameters
+    ----------
+    silent : bool
+        If true, defaults are chosen if no parameters found
+    client : str
+        The database URL to retrieve parameters for.  If None, the default is loaded
+
+    Returns
+    -------
+    A Params object for the AlyxClient
+    """
     client_key = _key_from_url(client) if client else None
     cache_map = iopar.read(f'{_PAR_ID_STR}/{_CLIENT_ID_STR}', {})
     if not cache_map:  # This can be removed in the future
@@ -137,9 +147,15 @@ def get(silent=False, client=None):
     return iopar.read(f'{_PAR_ID_STR}/{client_key or cache_map.DEFAULT}').set('CACHE_DIR', cache)
 
 
+def get_default_client():
+    """Returns the default AlyxClient URL, or None if no default is set."""
+    cache_map = iopar.as_dict(iopar.read(f'{_PAR_ID_STR}/{_CLIENT_ID_STR}', {})) or {}
+    return cache_map.get('DEFAULT', None)
+
+
 def save(par, client):
     # Remove cache dir variable before saving
-    par = {k: v for k, v in par.as_dict().items() if 'CACHE_DIR' not in k}
+    par = {k: v for k, v in iopar.as_dict(par).items() if 'CACHE_DIR' not in k}
     iopar.write(f'{_PAR_ID_STR}/{_key_from_url(client)}', par)
 
 
@@ -197,7 +213,7 @@ def _patch_params():
 
         # Save web client parameters
         new_web_client_pars = {k: v for k, v in old_par.as_dict().items()
-                               if k in default().as_dict()}
+                               if k in default().as_dict() or k == 'ALYX_LOGIN'}
         cache_name = _key_from_url(old_par.ALYX_URL)
         iopar.write(f'{_PAR_ID_STR}/{cache_name}', new_web_client_pars)